Added ExoPlayerModule
