+++ app/src/main/java/me/androidbox/busbybaking/recipeVideoSteps/RecipeVideoStepsView.java
@@ -19,6 +19,7 @@ import com.google.android.exoplayer2.trackselection.AdaptiveTrackSelection;
 import com.google.android.exoplayer2.trackselection.TrackSelector;
 import com.google.android.exoplayer2.ui.SimpleExoPlayerView;
 import com.google.android.exoplayer2.upstream.BandwidthMeter;
+import com.google.android.exoplayer2.util.Util;
 
 import org.parceler.Parcels;
