+++ build.gradle
@@ -10,8 +10,8 @@ buildscript {
         }
     }
     dependencies {
-    /*    classpath 'com.android.tools.build:gradle:3.0.0-alpha3'*/
-        classpath 'com.android.tools.build:gradle:2.3.2'
+        classpath 'com.android.tools.build:gradle:3.0.0-alpha3'
+        /*classpath 'com.android.tools.build:gradle:2.3.2'*/
         classpath 'com.dicedmelon.gradle:jacoco-android:0.1.1'
 
         // NOTE: Do not place your application dependencies here; they belong
